feat: defer user initialization with web worker (#1290)
* feat: init user with a web worker to defer work

* feat: defer authentication once app rendered
