+++ studio/src/app/types/core/ic.identity.ts
@@ -1,0 +1,4 @@
+export interface InternetIdentityAuth {
+  identityKey: string | null;
+  delegationChain: string | null;
+}
