+++ studio/src/app/utils/core/ic.identity.utils.ts
@@ -1,0 +1,24 @@
+import {Identity} from '@dfinity/agent';
+import {DelegationChain, DelegationIdentity, Ed25519KeyIdentity} from '@dfinity/identity';
+
+import {InternetIdentityAuth} from '../../types/core/ic.identity';
+import {LocalStorage} from '@dfinity/auth-client';
+
+export const internetIdentityAuth = async (): Promise<InternetIdentityAuth> => {
+  const storage: LocalStorage = new LocalStorage('ic-');
+
+  const identityKey: string | null = await storage.get('identity');
+  const delegationChain: string | null = await storage.get('delegation');
+
+  return {
+    identityKey,
+    delegationChain
+  };
+};
+
+export const initIdentity = ({identityKey, delegationChain}: {identityKey: string | null; delegationChain: string | null}): Identity => {
+  const chain: DelegationChain = DelegationChain.fromJSON(delegationChain);
+  const key: Ed25519KeyIdentity = Ed25519KeyIdentity.fromJSON(identityKey);
+
+  return DelegationIdentity.fromDelegation(key, chain);
+};
